# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Mausam 2.0 - Personalized Weather Intelligence",
  "description": "Mausam 2.0 - a Personalized Weather Intelligence Engine on official IMD/MoES data. Personalized homepage cards for the 'Mausam' app (SIH 2026 PS 26076).",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/alerts/simulate": {
   "post": {
    "tags": [
     "alerts"
    ],
    "summary": "Simulate",
    "operationId": "simulate_api_v1_alerts_simulate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AlertSimulationRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SimulationResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/admin/alerts": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Alerts",
    "operationId": "list_alerts_api_admin_alerts_get",
    "parameters": [
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create Alert",
    "operationId": "create_alert_api_admin_alerts_post",
    "parameters": [
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AlertCreate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/alerts/{alert_id}": {
   "put": {
    "tags": [
     "admin"
    ],
    "summary": "Update Alert",
    "operationId": "update_alert_api_admin_alerts__alert_id__put",
    "parameters": [
     {
      "name": "alert_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Alert Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AlertUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "admin"
    ],
    "summary": "Delete Alert",
    "operationId": "delete_alert_api_admin_alerts__alert_id__delete",
    "parameters": [
     {
      "name": "alert_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Alert Id"
      }
     },
     {
      "name": "mausam_admin_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mausam Admin Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AlertCreate": {
    "properties": {
     "severity": {
      "type": "string",
      "title": "Severity",
      "default": "orange"
     },
     "event_type": {
      "type": "string",
      "title": "Event Type",
      "default": "heavy_rain"
     },
     "headline": {
      "type": "string",
      "title": "Headline",
      "default": "Heavy rain expected"
     },
     "detail": {
      "type": "string",
      "title": "Detail",
      "default": ""
     },
     "region": {
      "type": "string",
      "title": "Region",
      "default": "Pune"
     },
     "impact": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Impact",
      "default": []
     },
     "valid_hours": {
      "type": "integer",
      "title": "Valid Hours",
      "default": 24
     },
     "radius_km": {
      "type": "number",
      "title": "Radius Km",
      "default": 40.0
     }
    },
    "type": "object",
    "title": "AlertCreate"
   },
...
   "AlertSimulationRequest": {
    "properties": {
     "severity": {
      "$ref": "#/components/schemas/WarningSeverity",
      "default": "orange"
     },
     "event_type": {
      "type": "string",
      "title": "Event Type",
      "default": "heavy_rain"
     },
     "headline": {
      "type": "string",
      "title": "Headline",
      "default": "Heavy rain expected"
     },
     "detail": {
      "type": "string",
      "title": "Detail",
      "default": ""
     },
     "region": {
      "type": "string",
      "title": "Region",
      "default": "Pune"
     },
     "impact": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Impact",
      "default": [
       "flooding on low-lying roads",
       "traffic disruption",
       "tree fall risk"
      ]
     },
     "valid_hours": {
      "type": "integer",
      "maximum": 168.0,
      "minimum": 1.0,
      "title": "Valid Hours",
      "default": 24
     },
     "center": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/GeoPoint"
       },
       {
        "type": "null"
       }
      ]
     },
     "radius_km": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 500.0,
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Radius Km"
     }
    },
    "type": "object",
    "title": "AlertSimulationRequest"
   },
...
   "AlertUpdate": {
    "properties": {
     "severity": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Severity"
     },
     "event_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Event Type"
     },
     "headline": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Headline"
     },
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detail"
     },
     "region": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Region"
     },
     "valid_hours": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Valid Hours"
     }
    },
    "type": "object",
    "title": "AlertUpdate"
   },
...
   "DisasterAlert": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "severity": {
      "$ref": "#/components/schemas/WarningSeverity"
     },
     "event_type": {
      "type": "string",
      "title": "Event Type"
     },
     "headline": {
      "type": "string",
      "title": "Headline"
     },
     "detail": {
      "type": "string",
      "title": "Detail",
      "default": ""
     },
     "region": {
      "type": "string",
      "title": "Region"
     },
     "polygon": {
      "$ref": "#/components/schemas/GeoPolygon"
     },
     "issued_by": {
      "type": "string",
      "title": "Issued By",
      "default": "IMD"
     },
     "issued_at": {
      "type": "string",
      "format": "date-time",
      "title": "Issued At"
     },
     "valid_until": {
      "type": "string",
      "format": "date-time",
      "title": "Valid Until"
     },
     "actionable": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Actionable",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "id",
     "severity",
     "event_type",
     "headline",
     "region",
     "polygon",
     "issued_at",
     "valid_until"
    ],
    "title": "DisasterAlert"
   },
...
   "GeoPoint": {
    "properties": {
     "lat": {
      "type": "number",
      "maximum": 90.0,
      "minimum": -90.0,
      "title": "Lat"
     },
     "lon": {
      "type": "number",
      "maximum": 180.0,
      "minimum": -180.0,
      "title": "Lon"
     },
     "geohash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Geohash"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "district": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "District"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     }
    },
    "type": "object",
    "required": [
     "lat",
     "lon"
    ],
    "title": "GeoPoint"
   },
   "GeoPolygon": {
    "properties": {
     "type": {
      "type": "string",
      "title": "Type",
      "default": "Polygon"
     },
     "coordinates": {
      "items": {
       "items": {
        "type": "number"
       },
       "type": "array"
      },
      "type": "array",
      "title": "Coordinates",
      "description": "ring: [[lon,lat],...]"
     }
    },
    "type": "object",
    "required": [
     "coordinates"
    ],
    "title": "GeoPolygon",
    "description": "GeoJSON-style polygon (ring of [lon, lat] pairs, closed)."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PushTarget": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "registered_city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Registered City"
     },
     "in_polygon": {
      "type": "boolean",
      "title": "In Polygon"
     },
     "headline": {
      "type": "string",
      "title": "Headline"
     },
     "body": {
      "type": "string",
      "title": "Body"
     },
     "fcm_payload": {
      "type": "object",
      "title": "Fcm Payload",
      "default": {}
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "registered_city",
     "in_polygon",
     "headline",
     "body"
    ],
    "title": "PushTarget"
   },
...
   "SimulationResult": {
    "properties": {
     "alert": {
      "$ref": "#/components/schemas/DisasterAlert"
     },
     "affected_users": {
      "items": {
       "$ref": "#/components/schemas/PushTarget"
      },
      "type": "array",
      "title": "Affected Users",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "alert"
    ],
    "title": "SimulationResult"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WarningSeverity": {
    "type": "string",
    "enum": [
     "green",
     "yellow",
     "orange",
     "red"
    ],
    "title": "WarningSeverity"
   },
...
  }
 }
}